+++ website/.vitepress/theme/landing/FolioHeader.vue
@@ -1,0 +1,16 @@
+<template>
+  <header v-pre class="folio" role="banner">
+    <div class="mark" aria-label="MemPalace">
+      <img src="/mempalace_logo.png" alt="" aria-hidden="true" />
+      <span>MemPalace</span>
+    </div>
+    <nav class="right" aria-label="Primary">
+      <a href="#anatomy" class="hide-mobile">Anatomy</a>
+      <a href="#dialect" class="hide-mobile">Dialect</a>
+      <a href="#mechanics" class="hide-mobile">Mechanics</a>
+      <a href="#install" class="hide-mobile">Install</a>
+      <a href="/guide/getting-started">Docs</a>
+      <a href="https://github.com/MemPalace/mempalace">GitHub ↗</a>
+    </nav>
+  </header>
+</template>
